pyproject.toml
[tool.poetry]
name = "proxy"
version = "0.1.0"
description = ""
authors = [
    "NC3 Team <info@nc3.lu>",
]
license = "AGPL-3.0-or-later"
readme = "README.md"

homepage = "https://github.com/informed-governance-project/proxy"
repository = "https://github.com/informed-governance-project/proxy"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
psycopg2-binary = "^2.8.4"
cryptography = "^39.0.1"
django-cors-headers = "^3.14.0"
drf-spectacular-sidecar = "^2023.4.1"
django-bootstrap5 = "^23.1"
django-otp = "^1.1.6"
django-formtools = "^2.4"
django-two-factor-auth = "^1.15.2"
phonenumbers = "^8.13.11"
django-revproxy = {url = "https://github.com/jazzband/django-revproxy/archive/refs/tags/0.11.0.tar.gz"}
django-import-export = "^3.2.0"
parler = "^1.0.1"
django-parler = "^2.3"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
pre-commit = "^2.20.0"
black = "^22.10.0"
pyupgrade = "^3.2.2"
autoflake = "^1.7.7"
mypy = "^0.991"
types-requests = "^2.28.11.5"
pip-audit = "^2.5.4"
isort = "^5.12.0"
django-debug-toolbar = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
strict_optional = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true

show_error_context = true
pretty = true
